refactor-splitted main reducer into smaller reducer fuction
