+++ src/reducer/index.js
@@ -1,0 +1,6 @@
+export {
+  filterReducer,
+  initialFilterState,
+} from './filter-reducer/filter-reducer';
+export { dataReducer, initialDataState } from './data-reducer/data-reducer';
+export { stateReducer } from './main-reducer/state-reducer';
